correção: Tipo da coluna numero_rua.
- Alterado de smallint para int

+++ database/migrations/2020_09_14_235821_create_pessoas_table.php
@@ -30,7 +30,7 @@ class CreatePessoasTable extends Migration
             $table->string('email');
 
             $table->string('rua');
-            $table->smallInteger('numero_rua');
+            $table->integer('numero_rua');
             $table->string('complemento')->nullable();
             $table->string('referencia')->nullable();
             $table->string('bairro');
